refactor hq
