--- segment_anything_hq/__init__.py
@@ -1,16 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .build_sam import (
-    build_sam,
-    build_sam_vit_h,
-    build_sam_vit_l,
-    build_sam_vit_b,
-    sam_model_registry,
-)
-from .build_sam_baseline import sam_model_registry_baseline
-from .predictor import SamPredictor
-from .automatic_mask_generator import SamAutomaticMaskGenerator
